lab8 passed

+++ proga/lab8/app/app/src/main/java/lab8/app/commands/CommandClear.java
@@ -1,11 +1,12 @@
 package lab8.app.commands;
 
+import java.util.List;
+import java.util.ResourceBundle;
 import java.util.concurrent.PriorityBlockingQueue;
 
 import lab8.app.database.DataBase;
 import lab8.app.labwork.LabWork;
 import lab8.app.signals.ServerSignal;
-import lab8.app.signals.Signal;
 
 
 /** Класс команды реализующей удаление всех элементов из коллекции.
@@ -14,21 +15,27 @@ import lab8.app.signals.Signal;
 public class CommandClear extends Command{
     
     public CommandClear() {
-        this.description = "Комманда Clear полностью очищает коллекцию элементов LabWork.";
     }
 
     @Override
-    public ServerSignal execute( PriorityBlockingQueue<LabWork> priorityBlockingQueue) {
+    public ServerSignal execute( PriorityBlockingQueue<LabWork> priorityBlockingQueue, ResourceBundle bundle) {
 
         ServerSignal resultSignal = new ServerSignal();
+
+        List<LabWork> labWorksToDelete = priorityBlockingQueue.stream().filter(labWork -> labWork.getOwner().getId() == getUser().getId()).toList();
+        if (labWorksToDelete.size() == 0) {
+            resultSignal.setMessage(bundle.getString("clearer1"));
+            resultSignal.setSucces(false);
+            return resultSignal;
+        }
 
         boolean cleared = DataBase.deleteAllLabWorks();
         if (cleared) {
             priorityBlockingQueue.clear();
-            resultSignal.setMessage("Коллекция была очищена.");
+            resultSignal.setMessage(bundle.getString("clearsuc"));
             resultSignal.setSucces(true);
         } else {
-            resultSignal.setMessage("Ошибка при очистке коллекции.");
+            resultSignal.setMessage(bundle.getString("clearer"));
             resultSignal.setSucces(false);
         }
