+++ proga/lab8/app/app/src/main/java/lab8/app/commands/CommandRemoveHead.java
@@ -1,11 +1,11 @@
 package lab8.app.commands;
 
+import java.util.ResourceBundle;
 import java.util.concurrent.PriorityBlockingQueue;
 
 import lab8.app.database.DataBase;
 import lab8.app.labwork.LabWork;
 import lab8.app.signals.ServerSignal;
-import lab8.app.signals.Signal;
 
 /** Класс команды реализующей удаление элемента в начале коллекции.
  * 
@@ -14,32 +14,37 @@ public class CommandRemoveHead extends Command {
 
 
     public CommandRemoveHead() {
-        this.description = "Команда Remvoe head удаляет самый старый элемент коллекции.";
     }
 
     LabWork operand;
     int index;
 
     @Override
-    public ServerSignal execute( PriorityBlockingQueue<LabWork> priorityBlockingQueue) {
+    public ServerSignal execute( PriorityBlockingQueue<LabWork> priorityBlockingQueue, ResourceBundle bundle) {
 
         ServerSignal resultSignal = new ServerSignal();
 
         if (priorityBlockingQueue.size() == 0) {
-            resultSignal.setMessage("Коллекция пуста.");
+            resultSignal.setMessage(bundle.getString("colempty"));
             resultSignal.setSucces(false);
             return resultSignal;
         }
         
         LabWork labWorkToDelete = priorityBlockingQueue.peek();
 
+        if (labWorkToDelete.getOwner().getId() != getUser().getId()) {
+            resultSignal.setMessage(bundle.getString("notowner"));
+            resultSignal.setSucces(false);
+            return resultSignal;
+        }
+
         boolean deleted = DataBase.deleteLabWorkById(labWorkToDelete.getId());
         if (deleted) {
             priorityBlockingQueue.poll();
-            resultSignal.setMessage("Верхний элемент удалён.");
+            resultSignal.setMessage(bundle.getString("headdel"));
             resultSignal.setSucces(true);
         } else {
-            resultSignal.setMessage("Верхний элемент не удалён по техническим причинам.");
+            resultSignal.setMessage(bundle.getString("elnotdel"));
             resultSignal.setSucces(false);
         }
